+++ Assets/Scripts/Network/RWGameProtocol/Msg/MsgHitDamage.cs
@@ -6,7 +6,7 @@ namespace RWGameProtocol.Msg
     public class MsgHitDamageReq
     {
         // 플레이어 UID
-        public int PlayerUId;
+        public ushort PlayerUId;
 
         // 데미지량. (실제 데미지량은 0.01을 곱해서 사용해야 함)
         public int Damage;
@@ -19,7 +19,7 @@ namespace RWGameProtocol.Msg
         public short ErrorCode;
 
         // 플레이어 UID
-        public int PlayerUId;
+        public ushort PlayerUId;
 
         // 데미지량 (실제 데미지량은 0.01을 곱해서 사용해야 함)
         public int Damage;
@@ -33,7 +33,7 @@ namespace RWGameProtocol.Msg
     public class MsgHitDamageNotify
     {
         // 플레이어 UID
-        public int PlayerUId;
+        public ushort PlayerUId;
 
         // 데미지량 (실제 데미지량은 0.01을 곱해서 사용해야 함)
         public int Damage;
